+++ pywren_ibm_cloud/storage/backends/ibm_cos/ibm_cos.py
@@ -67,7 +67,7 @@ class IBMCloudObjectStorageBackend:
             client_config = ibm_botocore.client.Config(max_pool_connections=128,
                                                        user_agent_extra=user_agent,
                                                        connect_timeout=1,
-                                                       read_timeout=5,
+                                                       read_timeout=3,
                                                        retries={'max_attempts': OBJ_REQ_RETRIES})
 
             self.cos_client = ibm_boto3.client('s3',
@@ -81,7 +81,7 @@ class IBMCloudObjectStorageBackend:
                                                        max_pool_connections=128,
                                                        user_agent_extra=user_agent,
                                                        connect_timeout=1,
-                                                       read_timeout=5,
+                                                       read_timeout=3,
                                                        retries={'max_attempts': OBJ_REQ_RETRIES})
 
             token_manager = DefaultTokenManager(api_key_id=api_key)